Update controller files to existing pattern
Separate provider and reconciler into separate files, move rbac
annotations to doc.go, some renames.
